Changes related with create hook for show modal for current order in each status for order details - useAdminModalOrder field
